requirements.txt
celery==4.4.6
Flask_Login==0.5.0
Flask_Avatars==0.2.2
SQLAlchemy==1.3.9
Flask==1.1.1
WTForms==2.2.1
pyquery==1.4.1
Flask_Script==2.0.6
alembic==1.4.0
Flask_SQLAlchemy==2.4.1
Faker==4.1.1
Flask_CLI==0.4.0
Flask_Migrate==2.5.3
feedparser==5.2.1
requests==2.22.0
Click==7.0
Flask_Mail==0.9.1
Flask_WTF==0.14.3
PyJWT==1.7.1
redis
